break out a component
